Removed invalid references to Appium and moved the Popupstore class to a seperate file
