+++ lib/helper/Puppeteer.js
@@ -11,76 +11,11 @@ const {
 } = require('../utils');
 const path = require('path');
 const ElementNotFound = require('./errors/ElementNotFound');
+const Popup = require('./extras/Popup');
 
 const puppeteer = requireg('puppeteer');
 
-/**
- * Class to handle the interaction with the Dialog (Popup) Class from Puppeteer
- */
-class PopupStore {
-  constructor(popup, defaultAction) {
-    this._popup = popup || {};
-    this._actionType = '';
-    this._defaultAction = defaultAction || '';
-  }
-
-  _assertValidActionType(action) {
-    if (['accept', 'cancel'].indexOf(action) === -1) {
-      throw new Error('Invalid Popup action type. Only "accept" or "cancel" actions are accepted');
-    }
-  }
-
-  set defaultAction(action) {
-    this._assertValidActionType(action);
-    this._defaultAction = action;
-  }
-
-  get defaultAction() {
-    return this._defaultAction;
-  }
-
-  get popup() {
-    return this._popup;
-  }
-
-  set popup(popup) {
-    if (this._popup) {
-      console.error('Popup already exists and was not closed. Popups must always be closed by calling either I.acceptPopup() or I.cancelPopup()');
-    }
-    this._popup = popup;
-  }
-
-  get actionType() {
-    return this._actionType;
-  }
-
-  set actionType(action) {
-    this._assertValidActionType(action);
-    this._actionType = action;
-  }
-
-  clear() {
-    this._popup = null;
-    this._actionType = '';
-  }
-
-  assertPopupVisible() {
-    if (!this._popup) {
-      throw new Error('There is no Popup visible');
-    }
-  }
-
-  assertPopupActionType(type) {
-    this.assertPopupVisible();
-    const expectedAction = this._actionType || this._defaultAction;
-    if (expectedAction !== type) {
-      throw new Error(`Popup action does not fit the expected action type. Expected popup action to be '${expectedAction}' not '${type}`);
-    }
-    this.clear();
-  }
-}
-
-const popupStore = new PopupStore();
+const popupStore = new Popup();
 
 /**
  * Uses [Google Chrome's Puppeteer](https://github.com/GoogleChrome/puppeteer) library to run tests inside headless Chrome.
@@ -182,7 +117,7 @@ class Puppeteer extends Helper {
   /**
    * Accepts the active JavaScript native popup window, as created by window.alert|window.confirm|window.prompt.
    * Don't confuse popups with modal windows, as created by [various
-   * libraries](http://jster.net/category/windows-modals-popups). Appium: support only web testing
+   * libraries](http://jster.net/category/windows-modals-popups).
    */
   acceptPopup() {
     popupStore.assertPopupActionType('accept');
@@ -204,7 +139,6 @@ class Puppeteer extends Helper {
 
   /**
    * Dismisses the active JavaScript popup, as created by window.alert|window.confirm|window.prompt.
-   * Appium: support only web testing
    */
   cancelPopup() {
     popupStore.assertPopupActionType('cancel');
@@ -212,7 +146,7 @@ class Puppeteer extends Helper {
 
   /**
    * Checks that the active JavaScript popup, as created by `window.alert|window.confirm|window.prompt`, contains the
-   * given string. Appium: support only web testing
+   * given string.
    */
   async seeInPopup(text) {
     popupStore.assertPopupVisible();
